Change the mba-related architecture
1. Remove copyright statements by intel requirement
2. Structure MbaInfo is moved to model/mba/mba.go
