+++ app/server.go
@@ -110,7 +110,7 @@ func Initialize(c *Config) (*restful.Container, error) {
 	policy := v1.PolicyResource{}
 	hospitality := v1.HospitalityResource{}
 	wls := v1.WorkLoadResource{Db: db}
-	mba := v1.MbaInfo{}
+	mba := v1.MbaResource{}
 
 	// Register controller to container
 	caches.Register(wsContainer)
