+++ lib/mba/mba.go
@@ -1,7 +1,3 @@
-// Copyright 2018 QCT (Quanta Cloud Technology). All rights reserved.
-// Use of this source code is governed by the Apache 2.0
-// license that can be found in the LICENSE file.
-
 package mba
 
 import (
